[FIX] website: fix failing standalone test with refreshed registry
Since [1] the `test_02_copy_ids_views_unlink_on_module_update`
standalone test fails because it cannot find the `is_seo_optimized`
field.

This happens because the `env`'s registry is changed when
`button_immediate_upgrade` is called: the types of the used records do
not belong to the correct registry after the first call.

This commit re-obtains the records used in the second part of the test
to avoid this issue.

[1]: https://github.com/odoo/odoo/commit/7f36a94a548f728aa89feb1a4facd59c91fc47bf

task-4422810
runbot-227670

closes odoo/odoo#223893

Signed-off-by: Chong Wang (cwg) <cwg@odoo.com>